backend/requirements.txt
# FastAPI et serveur ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données et ORM
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Celery et Redis
celery==5.3.4
redis==5.0.1
flower==2.0.1

# APScheduler pour tâches planifiées
apscheduler==3.10.4

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
shap==0.43.0

# Visualisation
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# HTTP clients et APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
tenacity==8.2.3

# Telegram Bot
python-telegram-bot==20.7

# Email
aiosmtplib==3.0.1

# Validation et parsing
email-validator==2.1.0
python-dateutil==2.8.2

# Monitoring et logs
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Utils
python-dotenv==1.0.0
typing-extensions==4.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
